Refactor how default script user is defined
Do not hard-code the default script user in the jinja file, and remove
the branching check from there. Instead use the same method as for
default script name.

This means we can write tests for the default user. With the older
method, we would need to write out a Dockerfile and then parse it back
in for testing.