Provide a schema for configuration YAML
Bundle a schema (Kwalify format) for the dogen configuration YAML
and use it to validate user supplied dogen configuration at run
time.

Add an extend_schema hook for plugins to adjust the schema to
reflect additional configuration the expect before the validation
runs.

Adjust the rpm plugin to extend the schema with a top-level key
'rpms' with a list of strings as the value. Since the schema for
the rpms: key is so simple, the schema data structure is inlined
in the plugin's extend_schema method.

Adjust the cct plugin to extend the schema with a more complex
definition of the cct configuration data structure. As it is more
complex, the schema is stored in a data file at dogen/plugins/cct
and loaded in by the plugin's extend_schema method.

The cct plugin's schema leaves the key 'configure' undefined. This
is configuration that is passed through to the cct tool. Either cct
should validate its own configuration or we should request a schema
from the cct tool to validate here (a future enhancement)

Adjust tests accordingly.