cct plugin: support for runtime changes
Add support to the cct plugin for runtime changes.

Adjust the template so that we set the CCT_MODULES_PATH environment
variable via Dockerfile ENV. This is then honoured by both the build
time and run time cct invocations.

Adjust the template so that any values in the list cct.runtime_changes
are added into the ENTRYPOINT command.

Extend the cct plugin so that a /cct/runtime key in the dogen YAML
is written out to a file cctruntime.yaml and cct.runtime_changes is
set so that it is included in the ENTRYPOINT command (using the scheme
described above)