Remove openshift key
The openshift key in the schema was used to conditionally switch a
USER statement in the template prior to the CMD running. It switched
it between "USER jboss" (wildfly/community specific) and "USER 185"
(openshift/jboss specific).

The same can be achieved with the newer "user:" key, which is also
more generic and useful for not wildfly/community/jboss/openshift.