sort repo_files when generating additional_repos
The output of glob.glob is not deterministic. Therefore, multiple runs
of dogen on the same source might generate slightly different outputs.
One example is the ordering of the "yum install" and "rm ..."
instructions when using additional repos.

Lexographically sort the list of repo files before populating
additional_repos to make this part deterministic.