Refactoring
The 'modules' section was reworked. A new 'repositories' section was
added (which replaced the top level 'dependencies') and another
'install' was added to be able to list all modules that need to be
"installed/processed" to generate the image.

Specifying `cache_url` in configuration file was moved from `artifact`
to `common` section.

For resources, if you specify only `path` to resource and at the same time
the `cache_url` will be set in configuration file - Concreate will still
try to fetch it from cacher, even if it was defined as 'local' path.

Dropped using 'dependencies' word across the code in favor of 'module repositories'.

Resource classes were moved to resource.py. Same happened to unit tests
for these classes.

Cleaned up logging across the tool.

The `prepare_external_repositories` method code from `tools.py`
was moved to generator.py and combined with `prepare_repositories`.

Handling of additional repository files is now done via resources too.

Signed-off-by: Marek Goldmann <mgoldman@redhat.com>