add support for generating help.md pages
Generate a help.md for all images based on a template file that we populate
with information from the input image.yaml.

See https://github.com/projectatomic/container-best-practices/blob/master/creating/help.adoc
for a description of the format, rationale etc. of help pages.  We differ a
little from the guidelines on the layout of our template.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>