Support defining help_template in image.yaml
I think this is likely to be the way this is most often used in
practice. An image author may wish to provide a custom help.md file,
potentially containing long-form static text, either directly in an
image.yaml, or indirectly via a module (perhaps to share with a
family of images).

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>