Move expose-services label logic out of template
Move the logic that calculates an io.openshift.expose-services
label out of the template.jinja and into the base generator class.

Only generate this label when building a Red Hat image (--redhat).

Do not generate the label if it has been statically defined in the
input image YAML already.

Do not add a port to the list in the expose-services label unless
a service is either provided or inferred via the socket library
(the service suffix is a mandatory part of the label specification)

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>