Introduce "docgen" command: module doc generation
This is an initial port of generate_docs.py from rcernich's fork
of cct_modules to a cekit command:

    https://github.com/rcernich/cct_module/blob/840739f604af2b3fdced61cf94d1626a4e059dbb/generate_docs.py

This command does not require an image descriptor (image.yaml) which
all the other commands do, therefore it hooks in at a different place
and as part of this initial implementation is effectively mutually
exclusive to those other commands. The UI around that needs some
tweaking.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>