docgen: Cut scan_for_modules
This was a function from the original script that looked for module
files under CWD in order to generate docs for them. The current
behaviour of "cekit docgen" is to generate docs for just one module
and assume it is in CWD directly.

Cut scan_for_modules, we may attempt to bring it back in a future
PR (if so, I'd redo the recursion with os.walk)

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>