Descriptor loading rewrite
Sharing the same code to read string representing paths and
JSON stings at the same time is not easy. To differentiate between these
we assume that provided paths are absolute. If we found the "/" character
in the beginning, we assume that this is a path, otherwise the text
is parsed.

Fixes #320