Add support for validating image sources
This adds a --validate switch to the build command. It interrupts the build process
early in the stage, just after the model is prepared and validated. This
is similar to --dry-run with the difference that --dry-run additionally generates
required files for the selected builder (which --validate does not).

This required redesigning build hooks which are now following:

1. prepare()
2. before_generate()
3. generate()
4. before_build()
5. run()

Fixes #551