Copy modules and artifacts when module is processed
This chnges the way we generate Dockerfile so that
artifacts and modules are not copied at the beginning of
the Dockerfile. This makes it possible to reuse
Docker (and other builder's) cache for content that
was not changed.

To make use of rapid development with cached you need to
make sure the module you modify is installed later in the process
because this invalidates cache and all modules installed
later will need to be executed again.

If you use Docker builder, use the `--no-squash` option.
Without this, the squash post processing will remove all
intermediate container images created at the original image build time
removing the cache at the same time.

Fixes #544