Add support for squashing for Buildah and Podman
This makes use of the squashing feature of the Buildah and
Podman builder. We squash images by default now.

It is possible to disable squashing by using the: `--no-squash`
switch for both: Podman and Buildah builders.

Additionally the `BUILDAH_LAYERS` environment variable documentation
was adeded for above builders.

Fixes #441
Fixes #597
Fixes #598