Deterministic git resource handling
Git resource handling uses now deterministic way for generating
target directories where these should be cloned.

If the 'target' key is defined, it will be used as-is. If the
'target' key is missing, the value of the 'name' key is used.
If the 'name' key is not used, then the target diretory is
computed based on the URL of the git repository. This will be the
base name of the URL, without any extensions (like '.git').

There should be no case where we have two module repositories
defined that have the same target directory. In such case this is
a mistake in configuration and proper 'name' or 'target' keys
should be set.

Fixes #621