Add support for 'FROM scratch'
This commit adds support for building container images
using the special base image 'scratch'. Using this feature
requires that image is prepared in smart way and not every
module will work because the base image is empty and does
not provide any libraries.

A common use case for this type of images is to use it as
metadata images or to store binary builds (potentially
being a result of mult-stage builds).

To support these use cases a new property was added to resources:
'dest'. This defines where the artifact should be placed within the image.

Fixes #487