Properly handle package managers in multi-stage builds
In case of multi-stage builds using different images with different
operating systems it is required to specify correct package
managers so that provided packages can be installed properly.

This commit fixes a bug where the package manager specification
was ignored in builder images and instead the target package manager
was used. Now, the specified package manager in the image will be used.

Fixes #646