Docs: Include all keys in top-most OSBS example
Include an example of every (sub) key for OSBS configuration in
the example YAML at the top of the page.

It was not clear enough to me where in the osbs heirarchy the
`koji_target` key belonged from the documentation alone, and I
had to double check by reading the source code.

I've ordered the keys to match the order of descriptions below.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>