Issue 779: Change how docker tags are interpreted and passed to behave tags
The docker image name (including tag) is passed directly to behave as the name of a test tag if the docker image name does not include a slash. Behave won't accept test tags that have a non-integer value after the colon. Given CEKit design, this means it is not possible to test an image that has a docker tag other than "latest" or something that can be coerced to an integer using python's `int` function. This changes that behaviour to never pass the docker tag on to behave (i.e anything after a colon).