test: behave: add --steps-ref command
Prior to this commit, when cloning the behave-test-steps URI,
CEKIt looked up the schema_version set in the descriptor in use,
defaulting to cekit.version.schema_version (which was usually 1);
prefixed a 'v' to get 'v1' and always switched to that ref ('v1').

Instead, Add a click.option such that we have self.params.steps_ref used
as the reference to check out, defaulting to 'v' +
cekit.version.schema_version.

Fixes #876.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>